WIP:Fix refreshWebServer() problem. Should be implemented in RTOS

+++ Core/Inc/DHT.h
@@ -10,8 +10,8 @@
 
 typedef struct
 {
-	float Temperature;
-	float Humidity;
+	uint8_t Temperature;
+	uint8_t Humidity;
 }DHT_DataTypedef;
